+++ frontend/src/App.js
@@ -1,7 +1,7 @@
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
-import MainPage from "./pages/MainPage";
-import LoginPage from "./pages/LoginPage";
-import SignupPage from "./pages/SignupPage";
+import MainPage from "./pages/MainPage/MainPage";
+import LoginPage from "./pages/LoginSignup/LoginPage";
+import SignupPage from "./pages/LoginSignup/SignupPage";
 import './App.css';
 
 function App() {
