+++ frontend/src/components/Header/Header.css
@@ -1,13 +1,3 @@
-/*heartbutton*/
-.button{
-    cursor:pointer;
-    transition: transform 300ms ease;
-    font-size: 100px;
-}
-.orange{
-  color:#FFBF87;
-}
-
 /*header*/
 .header{
     background-color: white;
