+++ frontend/src/components/common/Button/HeartButton/HeartButton.css
@@ -1,0 +1,9 @@
+/*heartbutton*/
+.button{
+    cursor:pointer;
+    transition: transform 300ms ease;
+    font-size: 100px;
+}
+.orange{
+  color:#FFBF87;
+}
