+++ frontend/src/components/placeCard/PlaceCard.css
