+++ frontend/src/components/searchBar/SearchBar.css
