+++ frontend/src/pages/LoginSignup/LoginPage.jsx
@@ -1,10 +1,10 @@
 import React, { useState } from "react";
 import { useNavigate } from "react-router-dom";
-import Header from "../components/Header";
-import LogInContainer from "../components/LogInContainer";
-import InputField from "../components/InputField";
-import Button from "../components/common/Button/ButtonLogin";
-import "../styles/LoginSignup.css"
+import Header from "../../components/Header/Header";
+import LogInContainer from "../../components/LoginSignup/LogInContainer";
+import InputField from "../../components/LoginSignup/InputField";
+import Button from "../../components/common/Button/ButtonLogin";
+import "./LoginSignup.css"
 
 const LoginPage = () => {
     const [email, setEmail] = useState("");
