+++ frontend/src/pages/MainPage/Main.style.jsx
